Updated FalconRoboticsTester as well as created four new autonomous routes

+++ FalconRoboticsTester/src/main/java/com/example/falconroboticstester/Vars.java
@@ -5,7 +5,7 @@ import com.acmerobotics.roadrunner.geometry.Pose2d;
 public class Vars{
 
     public int d1, d2, d3, d4;
-    public int c1, c2, c3, c4, c5, c6, c7, c8, c9, c10, c11, c12, c13, c14, c15, c16, c17, c18;
+    public double c1, c2, c3, c4, c5, c6, c7, c8, c9, c10, c11, c12, c13, c14, c15, c16, c17, c18;
 
     public Pose2d startPose;
     public Pose2d endPose;
